- Cambios para mejorar el procesamiento en paralelo. Si esta disponible ppl se usara.
- Los procesos de imagen devuelven el estado:
  /*
   * \brief Mensajes de estado del proceso
   */
  enum class Status
  {
    OK,                       // El proceso se ha ejecutado correctamente
    INCORRECT_INPUT_DATA,     // Error de los datos de entrada
    PROCESS_ERROR             // Error durante el proceso.
  };

+++ src/core/console.cpp
@@ -228,7 +228,7 @@ int CmdParameterOptions::getIndex(std::string value) const
 {
   for ( size_t i = 0; i < mOptions.size(); i++ ) {
     if (mOptions[i] == value) {
-      return i;
+      return static_cast<int>(i);
       break;
     }
   }
@@ -239,7 +239,7 @@ int CmdParameterOptions::getIndex() const
 {
   for ( size_t i = 0; i < mOptions.size(); i++ ) {
     if (mOptions[i] == mDefValue) {
-      return i;
+      return static_cast<int>(i);
       break;
     }
   }
